pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "A first-person dungeon crawler game"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
]